move level component to attribute package
